Changed mocha tests to specify lengths
By having multiple tests for a single prefix, I was able to pass the
check, however, I am guessing that the exercise wanted a specific test
for each specific case. Not sure if that is the case or if I need to
figure some things out.

It would seem rather tedious to test such basic things the way I did.
Would I be able to make more concise later on?